Pulling out PointLabel into common class
